+++ js/src/contracts/contracts.js
@@ -19,6 +19,7 @@ import Registry from './registry';
 import SignatureReg from './signaturereg';
 import TokenReg from './tokenreg';
 import GithubHint from './githubhint';
+import smsVerification from './sms-verification';
 
 let instance = null;
 
@@ -52,6 +53,10 @@ export default class Contracts {
 
   get githubHint () {
     return this._githubhint;
+  }
+
+  get smsVerification () {
+    return smsVerification;
   }
 
   static create (api) {
