+++ js/src/contracts/sms-verification.js
@@ -14,7 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-const checkIfRequested = (contract, account) => {
+import { stringify } from 'querystring';
+
+export const checkIfVerified = (contract, account) => {
+  return contract.instance.certified.call({}, [account]);
+};
+
+export const checkIfRequested = (contract, account) => {
   return new Promise((resolve, reject) => {
     contract.subscribe('Requested', {
       fromBlock: 0, toBlock: 'pending'
@@ -30,4 +36,17 @@ const checkIfRequested = (contract, account) => {
   });
 };
 
-export default checkIfRequested;
+export const postToServer = (query) => {
+  query = stringify(query);
+  return fetch('https://sms-verification.parity.io/?' + query, {
+    method: 'POST', mode: 'cors', cache: 'no-store'
+  })
+  .then((res) => {
+    return res.json().then((data) => {
+      if (res.ok) {
+        return data.message;
+      }
+      throw new Error(data.message || 'unknown error');
+    });
+  });
+};
