+++ js/src/util/check-if-tx-failed.js
